Apply Container/Presentational Pattern
