+++ components/character-window.tsx
@@ -1,10 +1,10 @@
 "use client"
 
-import { useState, useRef } from "react"
 import type { Character, Card, Equipment } from "../types"
 import { CharacterSlot } from "./character-slot"
 import { CharacterSearchModal } from "./ui/modal/CharacterSearchModal"
 import { useTranslations } from "next-intl"
+import { useCharacterWindow } from "../hooks/deck-builder/useCharacterWindow"
 
 interface CharacterWindowProps {
   selectedCharacters: number[]
@@ -47,76 +47,27 @@ export function CharacterWindow({
   awakening = {},
   onAwakeningSelect,
 }: CharacterWindowProps) {
-  const [showSelector, setShowSelector] = useState(false)
-  const [selectedSlot, setSelectedSlot] = useState<number>(-1)
-  const [searchTerm, setSearchTerm] = useState("")
-  const [sortBy, setSortBy] = useState<"name" | "rarity">("rarity")
-  const [sortDirection, setSortDirection] = useState<"desc" | "asc">("desc")
-  const modalRef = useRef<HTMLDivElement>(null)
-  const [slotHasExistingCharacter, setSlotHasExistingCharacter] = useState(false)
   const t = useTranslations()
-
-  const handleOpenSelector = (slot: number) => {
-    // 슬롯에 이미 캐릭터가 있는지 확인
-    const hasExistingCharacter = selectedCharacters[slot] !== -1
-    setSlotHasExistingCharacter(hasExistingCharacter)
-
-    // 슬롯 정보 저장 및 검  !== -1
-    setSlotHasExistingCharacter(hasExistingCharacter)
-
-    // 슬롯 정보 저장 및 검색 모달 열기
-    setSelectedSlot(slot)
-    setSearchTerm("")
-    setShowSelector(true)
-  }
-
-  // 캐릭터 선택 모달에서 -1(없음)을 선택했을 때 처리하는 로직 추가
-  const handleCharacterSelect = (characterId: number) => {
-    if (selectedSlot !== -1) {
-      // 슬롯에 이미 캐릭터가 있었다면 먼저 제거
-      if (slotHasExistingCharacter) {
-        onRemoveCharacter(selectedSlot)
-      }
-
-      // 새 캐릭터 추가 (characterId가 -1이면 슬롯을 비움)
-      if (characterId !== -1) {
-        onAddCharacter(characterId, selectedSlot)
-      }
-
-      setShowSelector(false)
-      setSelectedSlot(-1)
-      setSlotHasExistingCharacter(false)
-    }
-  }
-
-  // Filter out already selected characters to prevent duplicates
-  const availableForSelection = availableCharacters.filter((character) => !selectedCharacters.includes(character.id))
-
-  // Filter by search term
-  const filteredCharacters = availableForSelection.filter((character) =>
-    t(character.name).toLowerCase().includes(searchTerm.toLowerCase()),
-  )
-
-  // Update the sort function to respect direction
-  const sortedCharacters = [...filteredCharacters].sort((a, b) => {
-    let result = 0
-
-    if (sortBy === "name") {
-      result = t(a.name).localeCompare(t(b.name))
-    } else {
-      // Sort by rarity (UR > SSR > SR > R)
-      const rarityOrder = { UR: 4, SSR: 3, SR: 2, R: 1 }
-      result =
-        (rarityOrder[b.rarity as keyof typeof rarityOrder] || 0) -
-        (rarityOrder[a.rarity as keyof typeof rarityOrder] || 0)
-    }
-
-    // Apply sort direction
-    return sortDirection === "asc" ? -result : result
+  const {
+    showSelector,
+    searchTerm,
+    sortBy,
+    sortDirection,
+    sortedCharacters,
+    hasAnyCharacter,
+    handleOpenSelector,
+    handleCloseSelector,
+    handleCharacterSelect,
+    handleSearchChange,
+    handleSortByChange,
+    handleSortDirectionToggle,
+  } = useCharacterWindow({
+    selectedCharacters,
+    availableCharacters,
+    onAddCharacter,
+    onRemoveCharacter,
+    t,
   })
-
-  // Check if any character slot is filled
-  const hasAnyCharacter = selectedCharacters.some((id) => id !== -1)
 
   return (
     <div className="w-full">
@@ -149,22 +100,18 @@ export function CharacterWindow({
       {/* 새로운 캐릭터 검색 모달 컴포넌트 사용 */}
       <CharacterSearchModal
         isOpen={showSelector}
-        onClose={() => {
-          setShowSelector(false)
-          setSelectedSlot(-1)
-          setSlotHasExistingCharacter(false)
-        }}
+        onClose={handleCloseSelector}
         title={
           <h3 className="text-lg font-bold neon-text">
             {t("select_character") || "Select Character"}
           </h3>
         }
         searchTerm={searchTerm}
-        onSearchChange={setSearchTerm}
+        onSearchChange={handleSearchChange}
         sortBy={sortBy}
-        onSortByChange={(value) => setSortBy(value as "name" | "rarity")}
+        onSortByChange={(value) => handleSortByChange(value as "name" | "rarity")}
         sortDirection={sortDirection}
-        onSortDirectionChange={() => setSortDirection((prev) => (prev === "asc" ? "desc" : "asc"))}
+        onSortDirectionChange={handleSortDirectionToggle}
         sortOptions={[
           { value: "rarity", label: t("sort_by_rarity") || "Sort by Rarity" },
           { value: "name", label: t("sort_by_name") || "Sort by Name" },
@@ -172,21 +119,13 @@ export function CharacterWindow({
         searchPlaceholder={t("search_characters") || "Search characters"}
         characters={sortedCharacters}
         onSelectCharacter={handleCharacterSelect}
-        t={t}
         getCardInfo={getCardInfo}
         getSkill={getSkill}
         data={data}
         maxWidth="max-w-3xl"
         footer={
           <div className="flex justify-end">
-            <button
-              onClick={() => {
-                setShowSelector(false)
-                setSelectedSlot(-1)
-                setSlotHasExistingCharacter(false)
-              }}
-              className="neon-button px-4 py-2 rounded-lg text-sm"
-            >
+            <button onClick={handleCloseSelector} className="neon-button px-4 py-2 rounded-lg text-sm">
               {t("close") || "Close"}
             </button>
           </div>
